🎨 NicknameError 모델 분리
